changed size of id from 32Bit to 16Bit

+++ cop/include/Id.hpp
@@ -5,7 +5,7 @@
 
 namespace cop {
 
-    using ID_t = uint32_t;
+    using ID_t = uint16_t;
 
     template<ID_t id>
     class Id_t {
